cleanup merge

+++ IntroSkipper/Manager/QueueManager.cs
@@ -178,7 +178,6 @@ public class QueueManager(ILogger<QueueManager> logger, ILibraryManager libraryM
             _queuedEpisodes[seasonId] = seasonEpisodes;
         }
 
-        // Early return if already queued
         if (seasonEpisodes.Any(e => e.EpisodeId == episode.Id))
         {
             _logger.LogDebug(
@@ -217,9 +216,9 @@ public class QueueManager(ILogger<QueueManager> logger, ILibraryManager libraryM
             Name = episode.Name,
             IsAnime = isAnime,
             Path = episode.Path,
-            Duration = (int)duration,
-            IntroFingerprintEnd = (int)fingerprintDuration,
-            CreditsFingerprintStart = (int)(duration - maxCreditsDuration),
+            Duration = duration,
+            IntroFingerprintEnd = fingerprintDuration,
+            CreditsFingerprintStart = duration - maxCreditsDuration,
         });
 
         pluginInstance.TotalQueued++;
@@ -251,8 +250,8 @@ public class QueueManager(ILogger<QueueManager> logger, ILibraryManager libraryM
             EpisodeId = movie.Id,
             Name = movie.Name,
             Path = movie.Path,
-            Duration = (int)duration,
-            CreditsFingerprintStart = (int)(duration - pluginInstance.Configuration.MaximumMovieCreditsDuration),
+            Duration = duration,
+            CreditsFingerprintStart = duration - pluginInstance.Configuration.MaximumMovieCreditsDuration,
             IsMovie = true
         });
